fix: pool frame-level speaker mixes for Mix-LN
