+++ deployment/modules/fastspeech2.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from modules.commons.common_layers import NormalInitEmbedding as Embedding
-from modules.fastspeech.acoustic_encoder import FastSpeech2Acoustic
+from modules.fastspeech.acoustic_encoder import FastSpeech2Acoustic, uniform_attention_pooling
 from modules.fastspeech.variance_encoder import FastSpeech2Variance
 from utils.hparams import hparams
 from utils.phoneme_utils import PAD_INDEX
@@ -16,20 +16,6 @@ f0_max = 1100.0
 f0_min = 50.0
 f0_mel_min = 1127 * np.log(1 + f0_min / 700)
 f0_mel_max = 1127 * np.log(1 + f0_max / 700)
-
-
-def uniform_attention_pooling(spk_embed, durations):
-    _, T_mel, _ = spk_embed.shape
-    ph_starts = torch.cumsum(torch.cat([torch.zeros_like(durations[:, :1]), durations[:, :-1]], dim=1), dim=1)
-    ph_ends = ph_starts + durations
-    mel_indices = torch.arange(T_mel, device=spk_embed.device).view(1, 1, T_mel)
-    phoneme_to_mel_mask = (mel_indices >= ph_starts.unsqueeze(-1)) & (mel_indices < ph_ends.unsqueeze(-1))
-    uniform_scores = phoneme_to_mel_mask.float()
-    sum_scores = uniform_scores.sum(dim=2, keepdim=True)
-    attn_weights = uniform_scores / (sum_scores + (sum_scores == 0).float())  # [B, T_ph, T_mel]
-    ph_spk_embed = torch.bmm(attn_weights, spk_embed)
-
-    return ph_spk_embed
 
 
 def f0_to_coarse(f0):
