+++ modules/fastspeech/acoustic_encoder.py
@@ -12,6 +12,20 @@ from utils.hparams import hparams
 from utils.phoneme_utils import PAD_INDEX
 
 
+def uniform_attention_pooling(spk_embed, durations):
+    _, T_mel, _ = spk_embed.shape
+    ph_starts = torch.cumsum(torch.cat([torch.zeros_like(durations[:, :1]), durations[:, :-1]], dim=1), dim=1)
+    ph_ends = ph_starts + durations
+    mel_indices = torch.arange(T_mel, device=spk_embed.device).view(1, 1, T_mel)
+    phoneme_to_mel_mask = (mel_indices >= ph_starts.unsqueeze(-1)) & (mel_indices < ph_ends.unsqueeze(-1))
+    uniform_scores = phoneme_to_mel_mask.float()
+    sum_scores = uniform_scores.sum(dim=2, keepdim=True)
+    attn_weights = uniform_scores / (sum_scores + (sum_scores == 0).float())  # [B, T_ph, T_mel]
+    ph_spk_embed = torch.bmm(attn_weights, spk_embed)
+
+    return ph_spk_embed
+
+
 class FastSpeech2Acoustic(nn.Module):
     def __init__(self, vocab_size):
         super().__init__()
@@ -143,7 +157,11 @@ class FastSpeech2Acoustic(nn.Module):
             extra_embed = dur_embed + lang_embed
         else:
             extra_embed = dur_embed
-        encoder_out = self.encoder(txt_embed, extra_embed, txt_tokens == 0, spk_embed)
+        if self.use_mix_ln and spk_embed is not None and spk_embed.shape[1] > 1:
+            ph_spk_embed = uniform_attention_pooling(spk_embed, dur)
+        else:
+            ph_spk_embed = spk_embed
+        encoder_out = self.encoder(txt_embed, extra_embed, txt_tokens == 0, ph_spk_embed)
 
         encoder_out = F.pad(encoder_out, [0, 0, 1, 0])
         mel2ph_ = mel2ph[..., None].repeat([1, 1, encoder_out.shape[-1]])
